Modify recipe description, headings, and add lists
This recipe does not include the preparation process so I cited where readers can find the complete recipe in the description. The heading 'steps' was inaccurate so I changed it to be more specific to the process. The ingredients will match the outcome of the recipe. The lists include ingredients and steps.